Split Game off from App
